+++ hail/python/hail/experimental/sparse_mt/__init__.py
@@ -1,11 +1,7 @@
-from .vcf_combiner import run_combiner
 from .sparse_split_multi import sparse_split_multi
-from ...vds import lgt_to_gt
 from .densify import densify
 
 __all__ = [
-    'run_combiner',
     'sparse_split_multi',
-    'lgt_to_gt',
     'densify',
 ]
