+++ hail/python/hail/methods/__init__.py
@@ -74,8 +74,6 @@ __all__ = ['trio_matrix',
            'read_table',
            'get_vcf_metadata',
            'import_vcf',
-           'import_vcfs',
-           'import_gvcfs',
            'import_gvcf_interval',
            'index_bgen',
            'balding_nichols_model',
