+++ hail/python/hail/vds/combiner/combine.py
@@ -1,11 +1,15 @@
-from typing import Collection, List, Optional, Set, Tuple, Dict
+import math
+from typing import Collection, Optional, Set
+from typing import List, Tuple, Dict
 
 import hail as hl
 from hail import MatrixTable, Table
-from hail.ir import Apply, TableMapRows
-from hail.expr import construct_expr, unify_all
 from hail.experimental.function import Function
-from hail.experimental.vcf_combiner.vcf_combiner import combine_gvcfs, localize, parse_as_fields, unlocalize
+from hail.expr import StructExpression, unify_all, construct_expr
+from hail.expr.expressions import expr_bool, expr_str
+from hail.genetics.reference_genome import reference_genome_type
+from hail.ir import Apply, TableMapRows
+from hail.typecheck import oneof, sequenceof, typecheck
 from ..variant_dataset import VariantDataset
 
 _transform_variant_function_map: Dict[Tuple[hl.HailType, Tuple[str, ...]], Function] = {}
@@ -256,7 +260,7 @@ def make_reference_matrix_table(mt: MatrixTable,
                       hl.struct(END=row.info.END, **reference_fields, **handled_fields))
                 .or_error('found END with non reference-genotype at' + hl.str(row.locus)))
 
-    mt = localize(mt)
+    mt = localize(mt).key_by('locus')
     transform_row = _transform_reference_fuction_map.get((mt.row.dtype, entry_key))
     if transform_row is None or not hl.current_backend()._is_registered_ir_function_name(transform_row._name):
         transform_row = hl.experimental.define_function(
@@ -325,6 +329,7 @@ def combine_r(ts, ref_block_max_len_field):
         global_fds[ref_block_max_len_field] = hl.max(ts.g.map(lambda g: g[ref_block_max_len_field]))
     return ts.transmute_globals(**global_fds)
 
+
 def combine_reference_row(row, globals):
     merge_function = _merge_function_map.get((row.dtype, globals))
     if merge_function is None or not hl.current_backend()._is_registered_ir_function_name(merge_function._name):
@@ -346,6 +351,7 @@ def combine_reference_row(row, globals):
                      globals._ir)
     indices, aggs = unify_all(row, globals)
     return construct_expr(apply_ir, apply_ir.typ, indices, aggs)
+
 
 def combine_references(mts: List[MatrixTable]) -> MatrixTable:
     fd = hl.vds.VariantDataset.ref_block_max_length_field
@@ -370,3 +376,289 @@ def combine_variant_datasets(vdss: List[VariantDataset]) -> VariantDataset:
 
     variants = combine_gvcfs(no_variant_key)
     return VariantDataset(reference, variants._key_rows_by_assert_sorted('locus', 'alleles'))
+
+
+_transform_rows_function_map: Dict[Tuple[hl.HailType], Function] = {}
+_merge_function_map: Dict[Tuple[hl.HailType, hl.HailType], Function] = {}
+
+
+@typecheck(string=expr_str, has_non_ref=expr_bool)
+def parse_as_ints(string, has_non_ref):
+    ints = string.split(r'\|')
+    ints = hl.if_else(has_non_ref, ints[:-1], ints)
+    return ints.map(lambda i: hl.if_else((hl.len(i) == 0) | (i == '.'), hl.missing(hl.tint32), hl.int32(i)))
+
+
+@typecheck(string=expr_str, has_non_ref=expr_bool)
+def parse_as_doubles(string, has_non_ref):
+    ints = string.split(r'\|')
+    ints = hl.if_else(has_non_ref, ints[:-1], ints)
+    return ints.map(lambda i: hl.if_else((hl.len(i) == 0) | (i == '.'), hl.missing(hl.tfloat64), hl.float64(i)))
+
+
+@typecheck(string=expr_str, has_non_ref=expr_bool)
+def parse_as_sb_table(string, has_non_ref):
+    ints = string.split(r'\|')
+    ints = hl.if_else(has_non_ref, ints[:-1], ints)
+    return ints.map(lambda xs: xs.split(",").map(hl.int32))
+
+
+@typecheck(string=expr_str, has_non_ref=expr_bool)
+def parse_as_ranksum(string, has_non_ref):
+    typ = hl.ttuple(hl.tfloat64, hl.tint32)
+    items = string.split(r'\|')
+    items = hl.if_else(has_non_ref, items[:-1], items)
+    return items.map(lambda s: hl.if_else(
+        (hl.len(s) == 0) | (s == '.'),
+        hl.missing(typ),
+        hl.rbind(s.split(','), lambda ss: hl.if_else(
+            hl.len(ss) != 2,  # bad field, possibly 'NaN', just set it null
+            hl.missing(hl.ttuple(hl.tfloat64, hl.tint32)),
+            hl.tuple([hl.float64(ss[0]), hl.int32(ss[1])])))))
+
+
+_as_function_map = {
+    'AS_QUALapprox': parse_as_ints,
+    'AS_RAW_MQ': parse_as_doubles,
+    'AS_RAW_MQRankSum': parse_as_ranksum,
+    'AS_RAW_ReadPosRankSum': parse_as_ranksum,
+    'AS_SB_TABLE': parse_as_sb_table,
+    'AS_VarDP': parse_as_ints,
+}
+
+
+def parse_as_fields(info, has_non_ref):
+    return hl.struct(**{f: info[f] if f not in _as_function_map
+    else _as_function_map[f](info[f], has_non_ref) for f in info})
+
+
+def localize(mt):
+    if isinstance(mt, MatrixTable):
+        return mt._localize_entries('__entries', '__cols')
+    return mt
+
+
+def unlocalize(mt):
+    if isinstance(mt, Table):
+        return mt._unlocalize_entries('__entries', '__cols', ['s'])
+    return mt
+
+
+def merge_alleles(alleles):
+    from hail.expr.functions import _num_allele_type, _allele_ints
+    return hl.rbind(
+        alleles.map(lambda a: hl.or_else(a[0], ''))
+        .fold(lambda s, t: hl.if_else(hl.len(s) > hl.len(t), s, t), ''),
+        lambda ref:
+        hl.rbind(
+            alleles.map(
+                lambda al: hl.rbind(
+                    al[0],
+                    lambda r:
+                    hl.array([ref]).extend(
+                        al[1:].map(
+                            lambda a:
+                            hl.rbind(
+                                _num_allele_type(r, a),
+                                lambda at:
+                                hl.if_else(
+                                    (_allele_ints['SNP'] == at)
+                                    | (_allele_ints['Insertion'] == at)
+                                    | (_allele_ints['Deletion'] == at)
+                                    | (_allele_ints['MNP'] == at)
+                                    | (_allele_ints['Complex'] == at),
+                                    a + ref[hl.len(r):],
+                                    a)))))),
+            lambda lal:
+            hl.struct(
+                globl=hl.array([ref]).extend(hl.array(hl.set(hl.flatten(lal)).remove(ref))),
+                local=lal)))
+
+
+def combine_variant_rows(row, globals):
+    def renumber_entry(entry, old_to_new) -> StructExpression:
+        # global index of alternate (non-ref) alleles
+        return entry.annotate(LA=entry.LA.map(lambda lak: old_to_new[lak]))
+
+    merge_function = _merge_function_map.get((row.dtype, globals.dtype))
+    if merge_function is None or not hl.current_backend()._is_registered_ir_function_name(merge_function._name):
+        merge_function = hl.experimental.define_function(
+            lambda row, gbl:
+            hl.rbind(
+                merge_alleles(row.data.map(lambda d: d.alleles)),
+                lambda alleles:
+                hl.struct(
+                    locus=row.locus,
+                    alleles=alleles.globl,
+                    **({'rsid': hl.find(hl.is_defined, row.data.map(
+                        lambda d: d.rsid))} if 'rsid' in row.data.dtype.element_type else {}),
+                    __entries=hl.bind(
+                        lambda combined_allele_index:
+                        hl.range(0, hl.len(row.data)).flatmap(
+                            lambda i:
+                            hl.if_else(hl.is_missing(row.data[i].__entries),
+                                       hl.range(0, hl.len(gbl.g[i].__cols))
+                                       .map(lambda _: hl.missing(row.data[i].__entries.dtype.element_type)),
+                                       hl.bind(
+                                           lambda old_to_new: row.data[i].__entries.map(
+                                               lambda e: renumber_entry(e, old_to_new)),
+                                           hl.range(0, hl.len(alleles.local[i])).map(
+                                               lambda j: combined_allele_index[alleles.local[i][j]])))),
+                        hl.dict(hl.range(0, hl.len(alleles.globl)).map(
+                            lambda j: hl.tuple([alleles.globl[j], j])))))),
+            row.dtype, globals.dtype)
+        _merge_function_map[(row.dtype, globals.dtype)] = merge_function
+    indices, aggs = unify_all(row, globals)
+    apply_ir = Apply(merge_function._name,
+                     merge_function._ret_type,
+                     row._ir,
+                     globals._ir)
+    return construct_expr(apply_ir, apply_ir.typ, indices, aggs)
+
+
+def combine(ts):
+    ts = Table(TableMapRows(ts._tir, combine_variant_rows(
+        ts.row,
+        ts.globals)._ir))
+    return ts.transmute_globals(__cols=hl.flatten(ts.g.map(lambda g: g.__cols)))
+
+
+@typecheck(mts=sequenceof(oneof(Table, MatrixTable)))
+def combine_gvcfs(mts):
+    """Merges gvcfs and/or sparse matrix tables
+
+    Parameters
+    ----------
+    mts : :obj:`List[Union[Table, MatrixTable]]`
+        The matrix tables (or localized versions) to combine
+
+    Returns
+    -------
+    :class:`.MatrixTable`
+
+    Notes
+    -----
+    All of the input tables/matrix tables must have the same partitioning. This
+    module provides no method of repartitioning data.
+    """
+    ts = hl.Table.multi_way_zip_join([localize(mt) for mt in mts], 'data', 'g')
+    combined = combine(ts)
+    return unlocalize(combined)
+
+
+@typecheck(mt=hl.MatrixTable, desired_average_partition_size=int, tmp_path=str)
+def calculate_new_intervals(mt, desired_average_partition_size: int, tmp_path: str):
+    """takes a table, keyed by ['locus', ...] and produces a list of intervals suitable
+    for repartitioning a combiner matrix table.
+
+    Parameters
+    ----------
+    mt : :class:`.MatrixTable`
+        Sparse MT intermediate.
+    desired_average_partition_size : :obj:`int`
+        Average target number of rows for each partition.
+    tmp_path : :obj:`str`
+        Temporary path for scan checkpointing.
+
+    Returns
+    -------
+    (:obj:`List[Interval]`, :obj:`.Type`)
+    """
+    assert list(mt.row_key) == ['locus']
+    assert isinstance(mt.locus.dtype, hl.tlocus)
+    reference_genome = mt.locus.dtype.reference_genome
+    end = hl.Locus(reference_genome.contigs[-1],
+                   reference_genome.lengths[reference_genome.contigs[-1]],
+                   reference_genome=reference_genome)
+
+    (n_rows, n_cols) = mt.count()
+
+    if n_rows == 0:
+        raise ValueError('empty table!')
+
+    # split by a weight function that takes into account the number of
+    # dense entries per row. However, give each row some base weight
+    # to prevent densify computations from becoming unbalanced (these
+    # scale roughly linearly with N_ROW * N_COL)
+    ht = mt.select_rows(weight=hl.agg.count() + (n_cols // 25) + 1).rows().checkpoint(tmp_path)
+
+    total_weight = ht.aggregate(hl.agg.sum(ht.weight))
+    partition_weight = int(total_weight / (n_rows / desired_average_partition_size))
+
+    ht = ht.annotate(cumulative_weight=hl.scan.sum(ht.weight),
+                     last_weight=hl.scan._prev_nonnull(ht.weight),
+                     row_idx=hl.scan.count())
+
+    def partition_bound(x):
+        return x - (x % hl.int64(partition_weight))
+
+    at_partition_bound = partition_bound(ht.cumulative_weight) != partition_bound(ht.cumulative_weight - ht.last_weight)
+
+    ht = ht.filter(at_partition_bound | (ht.row_idx == n_rows - 1))
+    ht = ht.annotate(start=hl.or_else(
+        hl.scan._prev_nonnull(hl.locus_from_global_position(ht.locus.global_position() + 1,
+                                                            reference_genome=reference_genome)),
+        hl.locus_from_global_position(0, reference_genome=reference_genome)))
+    ht = ht.select(
+        interval=hl.interval(start=hl.struct(locus=ht.start), end=hl.struct(locus=ht.locus), includes_end=True))
+
+    intervals_dtype = hl.tarray(ht.interval.dtype)
+    intervals = ht.aggregate(hl.agg.collect(ht.interval))
+    last_st = hl.eval(
+        hl.locus_from_global_position(hl.literal(intervals[-1].end.locus).global_position() + 1,
+                                      reference_genome=reference_genome))
+    interval = hl.Interval(start=hl.Struct(locus=last_st), end=hl.Struct(locus=end), includes_end=True)
+    intervals.append(interval)
+    return intervals, intervals_dtype
+
+
+@typecheck(reference_genome=reference_genome_type, interval_size=int)
+def calculate_even_genome_partitioning(reference_genome, interval_size) -> List[hl.utils.Interval]:
+    """create a list of locus intervals suitable for importing and merging gvcfs.
+
+    Parameters
+    ----------
+    reference_genome: :class:`str` or :class:`.ReferenceGenome`,
+        Reference genome to use. NOTE: only GRCh37 and GRCh38 references
+        are supported.
+    interval_size: :obj:`int` The ceiling and rough target of interval size.
+        Intervals will never be larger than this, but may be smaller.
+
+    Returns
+    -------
+    :obj:`List[Interval]`
+    """
+
+    def calc_parts(contig):
+        def locus_interval(start, end):
+            return hl.Interval(
+                start=hl.Locus(contig=contig, position=start, reference_genome=reference_genome),
+                end=hl.Locus(contig=contig, position=end, reference_genome=reference_genome),
+                includes_end=True)
+
+        contig_length = reference_genome.lengths[contig]
+        n_parts = math.ceil(contig_length / interval_size)
+        real_size = math.ceil(contig_length / n_parts)
+        n = 1
+        intervals = []
+        while n < contig_length:
+            start = n
+            end = min(n + real_size, contig_length)
+            intervals.append(locus_interval(start, end))
+            n = end + 1
+
+        return intervals
+
+    if reference_genome.name == 'GRCh37':
+        contigs = [f'{i}' for i in range(1, 23)] + ['X', 'Y', 'MT']
+    elif reference_genome.name == 'GRCh38':
+        contigs = [f'chr{i}' for i in range(1, 23)] + ['chrX', 'chrY', 'chrM']
+    else:
+        raise ValueError(
+            f"Unsupported reference genome '{reference_genome.name}', "
+            "only 'GRCh37' and 'GRCh38' are supported")
+
+    intervals = []
+    for ctg in contigs:
+        intervals.extend(calc_parts(ctg))
+    return intervals
