+++ hail/python/test/hail/vds/test_combiner.py
@@ -28,35 +28,30 @@ def test_combiner_works():
     _paths = ['gvcfs/HG00096.g.vcf.gz', 'gvcfs/HG00268.g.vcf.gz']
     paths = [resource(p) for p in _paths]
     parts = [
-        hl.Interval(start=hl.Struct(locus=hl.Locus('chr20', 17821257, reference_genome='GRCh38')),
-                    end=hl.Struct(locus=hl.Locus('chr20', 18708366, reference_genome='GRCh38')),
+        hl.Interval(start=hl.Locus('chr20', 17821257, reference_genome='GRCh38'),
+                    end=hl.Locus('chr20', 18708366, reference_genome='GRCh38'),
                     includes_end=True),
-        hl.Interval(start=hl.Struct(locus=hl.Locus('chr20', 18708367, reference_genome='GRCh38')),
-                    end=hl.Struct(locus=hl.Locus('chr20', 19776611, reference_genome='GRCh38')),
+        hl.Interval(start=hl.Locus('chr20', 18708367, reference_genome='GRCh38'),
+                    end=hl.Locus('chr20', 19776611, reference_genome='GRCh38'),
                     includes_end=True),
-        hl.Interval(start=hl.Struct(locus=hl.Locus('chr20', 19776612, reference_genome='GRCh38')),
-                    end=hl.Struct(locus=hl.Locus('chr20', 21144633, reference_genome='GRCh38')),
+        hl.Interval(start=hl.Locus('chr20', 19776612, reference_genome='GRCh38'),
+                    end=hl.Locus('chr20', 21144633, reference_genome='GRCh38'),
                     includes_end=True)
     ]
-    vcfs = hl.import_gvcfs(paths, parts, reference_genome='GRCh38', array_elements_required=False)
-    entry_to_keep = defined_entry_fields(vcfs[0].filter_rows(hl.is_defined(vcfs[0].info.END)), 100_000) - {'GT', 'PGT', 'PL'}
-    vcfs = [transform_gvcf(mt.annotate_rows(info=mt.info.annotate(
-        MQ_DP=hl.missing(hl.tint32),
-        VarDP=hl.missing(hl.tint32),
-        QUALapprox=hl.missing(hl.tint32))),
-                           reference_entry_fields_to_keep=entry_to_keep)
-            for mt in vcfs]
-    comb = combine_variant_datasets(vcfs)
-    assert len(parts) == comb.variant_data.n_partitions()
-    comb.variant_data._force_count_rows()
-    comb.reference_data._force_count_rows()
+    with hl.TemporaryDirectory() as tmpdir:
+        out = os.path.join(tmpdir, 'out.vds')
+        hl.vds.new_combiner(temp_path=tmpdir, output_path=out, gvcf_paths=paths, intervals=parts, reference_genome='GRCh38').run()
+        comb = hl.vds.read_vds(out)
+        assert len(parts) == comb.variant_data.n_partitions()
+        comb.variant_data._force_count_rows()
+        comb.reference_data._force_count_rows()
 
 
 @fails_local_backend
 @fails_service_backend
 @test_timeout(12 * 60)
 def test_vcf_vds_combiner_equivalence():
-    import hail.experimental.vcf_combiner.vcf_combiner as vcf
+    import hail.experimental.sparse_mt.vcf_combiner as vcf
     import hail.vds.combiner as vds
     _paths = ['gvcfs/HG00096.g.vcf.gz', 'gvcfs/HG00268.g.vcf.gz']
     paths = [resource(p) for p in _paths]
